Refactored database_class3.py into separate files
Split database_class3.py into separate files for the Database and Gui portions, and for the shared functions
Added Global Variables and accessors into shared_functions.py
Updated todo list
