+++ todo.md
@@ -8,6 +8,7 @@ Add ability to update database without restarting script
 ## Refactoring
 
 Split differing functionalities into different scripts
+Strip scripts of all superfluous code
 
 ## Tests
